Pin: rename feature pin to lockscreen
