--- features/pin/api/src/main/kotlin/io/element/android/features/pin/api/PinStateService.kt
@@ -1,27 +1,0 @@
-/*
- * Copyright (c) 2023 New Vector Ltd
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package io.element.android.features.pin.api
-
-import kotlinx.coroutines.flow.StateFlow
-
-interface PinStateService {
-    val pinState: StateFlow<PinState>
-
-    suspend fun entersForeground()
-    suspend fun entersBackground()
-    suspend fun unlock()
-}
