--- features/pin/impl/src/main/kotlin/io/element/android/features/pin/impl/presentation/PinFlowNode.kt
@@ -1,78 +1,0 @@
-/*
- * Copyright (c) 2023 New Vector Ltd
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package io.element.android.features.pin.impl.presentation
-
-import android.os.Parcelable
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import com.bumble.appyx.core.composable.Children
-import com.bumble.appyx.core.modality.BuildContext
-import com.bumble.appyx.core.node.Node
-import com.bumble.appyx.core.plugin.Plugin
-import com.bumble.appyx.navmodel.backstack.BackStack
-import dagger.assisted.Assisted
-import dagger.assisted.AssistedInject
-import io.element.android.anvilannotations.ContributesNode
-import io.element.android.features.pin.impl.presentation.auth.PinAuthenticationNode
-import io.element.android.features.pin.impl.presentation.create.CreatePinNode
-import io.element.android.libraries.architecture.BackstackNode
-import io.element.android.libraries.architecture.animation.rememberDefaultTransitionHandler
-import io.element.android.libraries.architecture.createNode
-import io.element.android.libraries.di.AppScope
-import kotlinx.parcelize.Parcelize
-
-@ContributesNode(AppScope::class)
-class PinFlowNode @AssistedInject constructor(
-    @Assisted buildContext: BuildContext,
-    @Assisted plugins: List<Plugin>,
-) : BackstackNode<PinFlowNode.NavTarget>(
-    backstack = BackStack(
-        initialElement = NavTarget.Auth,
-        savedStateMap = buildContext.savedStateMap,
-    ),
-    buildContext = buildContext,
-    plugins = plugins,
-) {
-
-    sealed interface NavTarget : Parcelable {
-        @Parcelize
-        data object Auth : NavTarget
-
-        @Parcelize
-        data object Create : NavTarget
-    }
-
-    override fun resolve(navTarget: NavTarget, buildContext: BuildContext): Node {
-        return when (navTarget) {
-            NavTarget.Auth -> {
-                createNode<PinAuthenticationNode>(buildContext)
-            }
-            NavTarget.Create -> {
-                createNode<CreatePinNode>(buildContext)
-            }
-        }
-    }
-
-    @Composable
-    override fun View(modifier: Modifier) {
-        Children(
-            navModel = backstack,
-            modifier = modifier,
-            transitionHandler = rememberDefaultTransitionHandler(),
-        )
-    }
-}
